chore: Move pipeline group types to their own file
